feat(vibrant-core): add overflow systemProp
