+++ packages/vibrant-core/src/lib/props/display/display.ts
@@ -45,34 +45,4 @@ const hiddenProp = createSystemProp({
       : {},
 });
 
-const overflowXProp = createSystemProp({
-  property: 'overflowX',
-});
-
-const overflowYProp = createSystemProp({
-  property: 'overflowY',
-});
-
-const hideScrollProp = createSystemProp({
-  property: 'hideScroll',
-  transform: (value: boolean) =>
-    value
-      ? {
-          scrollbarWidth: 'none',
-          msOverflowStyle: 'none',
-          '&::-webkit-scrollbar': {
-            display: 'none',
-          },
-        }
-      : {},
-});
-
-export const displayProps = [
-  displayProp,
-  visibilityProp,
-  opacityProp,
-  hiddenProp,
-  overflowXProp,
-  overflowYProp,
-  hideScrollProp,
-];
+export const displayProps = [displayProp, visibilityProp, opacityProp, hiddenProp];
