+++ packages/vibrant-core/src/lib/props/index.ts
@@ -7,6 +7,7 @@ export type {
   FlexboxProps,
   InputProps,
   InteractionProps,
+  OverflowProps,
   PositionProps,
   PseudoClassProps,
   SizingProps,
