+++ packages/vibrant-core/src/lib/props/overflow/index.ts
@@ -1,0 +1,2 @@
+export type { OverflowProps } from './type';
+export { overflowProps } from './overflow';
