+++ packages/vibrant-core/src/lib/props/overflow/overflow.ts
@@ -1,0 +1,25 @@
+import { createSystemProp } from '../../createSystemProp';
+
+const overflowXProp = createSystemProp({
+  property: 'overflowX',
+});
+
+const overflowYProp = createSystemProp({
+  property: 'overflowY',
+});
+
+const hideScrollProp = createSystemProp({
+  property: 'hideScroll',
+  transform: (value: boolean) =>
+    value
+      ? {
+          scrollbarWidth: 'none',
+          msOverflowStyle: 'none',
+          '&::-webkit-scrollbar': {
+            display: 'none',
+          },
+        }
+      : {},
+});
+
+export const overflowProps = [overflowXProp, overflowYProp, hideScrollProp];
