+++ packages/vibrant-core/src/lib/props/overflow/type.ts
@@ -1,0 +1,7 @@
+import type { ResponsiveValue } from '../../../types';
+
+export type OverflowProps = {
+  overflowX?: ResponsiveValue<'hidden' | 'scroll' | 'visible'>;
+  overflowY?: ResponsiveValue<'hidden' | 'scroll' | 'visible'>;
+  hideScroll?: ResponsiveValue<boolean>;
+};
